Contextの差し替え処理を別ファイルに分けた
